requirements.txt
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-auth==2.23.4
google-cloud-storage==2.10.0

# Machine Learning Enhancements
tensorflow>=2.10.0
keras>=2.10.0

# Existing packages (if you have them)
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
dlib>=19.22.0
ultralytics>=8.0.0
flask>=2.0.0
requests>=2.25.0
stripe>=7.0.0
python-dotenv>=1.0.0gunicorn>=21.2.0
flask-cors>=3.0.10
